// src/app/components/Image/Image.generated.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
const defaultOptions = {} as const;
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** country code */
  CountryCode: any;
  /** An RFC-3339 compliant Full Date Scalar */
  Date: any;
  /** An RFC-3339 compliant DateTime Scalar */
  DateTime: any;
  /** A field whose value conforms to the standard internet email address format as specified in RFC822: https://www.w3.org/Protocols/rfc822/. */
  EmailAddress: any;
  /** gender */
  Gender: any;
  /** Long type */
  Long: any;
};
// ...
export type AutoQualityInput = {
  allowedError?: InputMaybe<Scalars['Float']>;
  maxFileOther?: InputMaybe<Scalars['Float']>;
  maxFileSize?: InputMaybe<Scalars['Int']>;
  maxQuality?: InputMaybe<Scalars['Int']>;
  method?: InputMaybe<ImgOptimazeType>;
  minQuality?: InputMaybe<Scalars['Int']>;
};
// ...
export type ImageTransformInput = {
  autoQuality?: InputMaybe<AutoQualityInput>;
  dpr?: InputMaybe<Scalars['Float']>;
  enlarge?: InputMaybe<Scalars['Boolean']>;
  extend?: InputMaybe<Scalars['Boolean']>;
  height?: InputMaybe<Scalars['Int']>;
  quality?: InputMaybe<Scalars['Int']>;
  resizeType?: InputMaybe<ImgResizeType>;
  width?: InputMaybe<Scalars['Int']>;
};

export enum ImgOptimazeType {
  Dssim = 'DSSIM',
  Ml = 'ML',
  Size = 'SIZE'
}

export enum ImgResizeType {
  Auto = 'AUTO',
  Fill = 'FILL',
  FillDown = 'FILL_DOWN',
  Fit = 'FIT',
  Force = 'FORCE'
}
// ...
export type AppRouterOptimizedImageQueryVariables = Types.Exact<{
  url: Types.Scalars['String'];
  transform?: Types.InputMaybe<Types.ImageTransformInput>;
}>;


export type AppRouterOptimizedImageQuery = { transformImage: { id: string, fileName: string, optimized?: string | null } };


export const AppRouterOptimizedImageDocument = gql`
    query AppRouterOptimizedImage($url: String!, $transform: ImageTransformInput) {
  transformImage: imgTransformImage(url: $url) {
    id
    optimized: imgSrc(transform: $transform)
    fileName
  }
}
    `;